+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - `app:bookmark_set_int(name, value)` lua method to set integer value of a bookmark
 - `app:region_pattern_fill(name, pattern)` lua method to fill a region
 - Context menu to copy bookmark names in bookmarks window
+- Make the offsets in the find dialog copiable/pasteable
 
 ### Changed